backend/pyproject.toml
[project]
name = "equiguard-backend"
version = "0.1.0"
description = "FastAPI backend for EquiGuard horse detection API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "sqlalchemy==2.0.43",
    "pydantic==2.11.7",
    "python-multipart==0.0.20",
    "Pillow==11.3.0",
    "langchain==0.3.27",
    "langchain-core==0.3.75",
    "langchain-anthropic==0.3.19",
    "langchain-openai==0.3.31",
    "langchain-ollama==0.3.6",
    "sentence-transformers>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
